model_doc_agent/requirements.txt
openai
tiktoken
PyYAML
pytest
langchain
langchain-openai
langchain-community
python-dotenv
langchain-anthropic 